[Refactor] rm large int value
