+++ be/src/column/fixed_length_column_base.cpp
@@ -16,6 +16,7 @@
 
 #include "base/simd/gather.h"
 #include "base/types/decimal12.h"
+#include "base/types/int128.h"
 #include "base/types/int256.h"
 #include "column/column_helper.h"
 #include "column/vectorized_fwd.h"
@@ -24,7 +25,6 @@
 #include "gutil/casts.h"
 #include "gutil/strings/fastmem.h"
 #include "gutil/strings/substitute.h"
-#include "types/large_int_value.h"
 #include "util/hash_util.hpp"
 #include "util/mysql_row_buffer.h"
 #include "util/value_generator.h"
