+++ be/src/exprs/cast_expr.h
@@ -17,6 +17,7 @@
 #include <type_traits>
 #include <utility>
 
+#include "base/types/int128.h"
 #include "column/array_column.h"
 #include "column/chunk.h"
 #include "column/column_helper.h"
@@ -26,7 +27,6 @@
 #include "exprs/expr.h"
 #include "jsonpath.h"
 #include "runtime/types.h"
-#include "types/large_int_value.h"
 #include "variant_path_parser.h"
 
 namespace starrocks {
@@ -404,7 +404,7 @@ struct CastToString {
             return v.to_string();
         } else if constexpr (IsInt128<Type>) {
             // int128_t
-            return LargeIntValue::to_string(v);
+            return int128_to_string(v);
         } else if constexpr (IsInt256<Type>) {
             // int256_t
             return v.to_string();
