+++ be/src/storage/types.cpp
@@ -38,6 +38,7 @@
 #include "base/string/slice.h"
 #include "base/string/string_parser.hpp"
 #include "base/types/decimal12.h"
+#include "base/types/int128.h"
 #include "base/utility/mem_util.hpp"
 #include "base/utility/mysql_global.h"
 #include "gutil/strings/numbers.h"
@@ -55,7 +56,6 @@
 #include "storage/type_traits.h"
 #include "types/array_type_info.h"
 #include "types/date_value.hpp"
-#include "types/large_int_value.h"
 #include "types/map_type_info.h"
 #include "types/struct_type_info.h"
 #include "util/hash_util.hpp"
