+++ be/test/exprs/string_fn_field.cpp
@@ -20,12 +20,12 @@
 
 #include "base/testutil/assert.h"
 #include "base/testutil/parallel_test.h"
+#include "base/types/int128.h"
 #include "column/array_column.h"
 #include "exprs/function_helper.h"
 #include "exprs/mock_vectorized_expr.h"
 #include "exprs/string_functions.h"
 #include "runtime/types.h"
-#include "types/large_int_value.h"
 
 namespace starrocks {
