--- be/test/runtime/large_int_value_test.cpp
@@ -1,68 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-#include "types/large_int_value.h"
-
-#include <gtest/gtest.h>
-
-#include <sstream>
-#include <string>
-
-#include "common/configbase.h"
-#include "types/constexpr.h"
-
-namespace starrocks {
-
-class LargeIntValueTest : public testing::Test {
-public:
-    LargeIntValueTest() = default;
-
-protected:
-    void SetUp() override {}
-    void TearDown() override {}
-};
-
-TEST_F(LargeIntValueTest, string_to_largeint) {
-    {
-        std::string str("1024");
-        std::stringstream ss;
-        ss << str;
-        __int128 v;
-        ss >> v;
-        ASSERT_EQ(v, 1024);
-    }
-
-    {
-        std::string str("170141183460469231731687303715884105727");
-        std::stringstream ss;
-        ss << str;
-        __int128 v;
-        ss >> v;
-        ASSERT_TRUE(v == MAX_INT128);
-    }
-
-    {
-        std::string str("-170141183460469231731687303715884105728");
-        std::stringstream ss;
-        ss << str;
-        __int128 v;
-        ss >> v;
-        ASSERT_TRUE(v == MIN_INT128);
-    }
-}
-
-} // end namespace starrocks
